requirements.txt
aiofiles==24.1.0
aiosmtplib==3.0.2
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
bcrypt==4.0.1
blinker==1.9.0
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
cssselect2==0.7.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.7
fastapi-mail==1.4.2
fonttools==4.55.8
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
loguru==0.7.3
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==24.2
passlib==1.7.4
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydyf==0.11.0
Pygments==2.18.0
PyJWT==2.10.1
pyphen==0.17.2
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.12.0
setuptools==75.8.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typer==0.15.1
typing_extensions==4.12.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
weasyprint==63.1
webencodings==0.5.1
websockets==14.1
wheel==0.45.1
zopfli==0.2.3.post1
